requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.15
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-logentry-admin==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.0
pyfiglet
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
simplejson==3.19.2
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
setuptools
requests
pyinstaller
psycopg2-binary
django-simple-history
django-jazzmin-admin-rangefilter

pylint-django
whitenoise
pycups